+++ share/pot/clone-fscomp.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
+:
 
+# shellcheck disable=SC2039
 clone-fscomp-help()
 {
 	echo "pot clone-fscomp [-hv] -f fscomp -F fscomp"
@@ -13,26 +15,28 @@ clone-fscomp-help()
 # $2 old fscomp name
 _cf_zfs()
 {
-	local _fscomp _cfscomp _fsdset _pdir _pbdir
-	local _node _mnt_p _opt _new_mnt_p
+	# shellcheck disable=SC2039
+	local _fscomp _cfscomp _fsdset _fsdir _snap
 	_fscomp=$1
 	_cfscomp=$2
 	_fsdset=${POT_ZFS_ROOT}/fscomp
 	_fsdir=${POT_FS_ROOT}/fscomp
-	_snap=$( _zfs_last_snap $_fsdset/$_cfscomp )
+	_snap=$( _zfs_last_snap "$_fsdset/$_cfscomp" )
 	if [ -z "$_snap" ]; then
 		_error "$_fsdset/$_cfscomp has no snapshots - please take one"
 		return 1
 	else
-		_debug "Cloning $_cfscompt@$_snap into $_fsdset/$_fscomp"
-		zfs clone -o mountpoint=$_fsdir/$_fscomp $_fsdset/$_cfscomp@$_snap $_fsdset/$_fscomp
+		_debug "Cloning $_cfscomp@$_snap into $_fsdset/$_fscomp"
+		zfs clone -o mountpoint="$_fsdir/$_fscomp" "$_fsdset/$_cfscomp@$_snap" "$_fsdset/$_fscomp"
 	fi
 	return 0 # true
 }
 
+# shellcheck disable=SC2039
 pot-clone-fscomp()
 {
-	local _pname _ipaddr _potbase _pb_ipaddr
+	# shellcheck disable=SC2039
+	local _fscomp _cfscomp
 	_fscomp=
 	_cfscomp=
 	OPTIND=1
@@ -68,18 +72,18 @@ pot-clone-fscomp()
 		clone-fscomp-help
 		${EXIT} 1
 	fi
-	if _zfs_dataset_valid ${POT_ZFS_ROOT}/fscomp/$_fscomp ; then
+	if _zfs_dataset_valid "${POT_ZFS_ROOT}/fscomp/$_fscomp" ; then
 		_error "fscomp $_fscomp already exists"
 		${EXIT} 1
 	fi
-	if ! _zfs_dataset_valid ${POT_ZFS_ROOT}/fscomp/$_cfscomp ; then
+	if ! _zfs_dataset_valid "${POT_ZFS_ROOT}/fscomp/$_cfscomp" ; then
 		_error "fscomp $_cfscomp doesn't exist"
 		${EXIT} 1
 	fi
 	if ! _is_uid0 ; then
 		${EXIT} 1
 	fi
-	if ! _cf_zfs $_fscomp $_cfscomp ; then
+	if ! _cf_zfs "$_fscomp" "$_cfscomp" ; then
 		${EXIT} 1
 	fi
 }